Fix export of fp8 ONNX files (#52)
* Fix export of fp8 ONNX files


* test: add fp8 calibration dtype regression test for #52

`_collect_calibration_dets` returns uint8; casting to float32 before
passing to mq.quantize triggered an INVALID_ARGUMENT error from the
ONNX runtime ("expected: tensor(uint8), got: tensor(float)").
The new test mirrors the existing int8 variant and asserts that the
fp8 path preserves the original uint8 dtype and forwards the
FP8-specific kwargs (op_types_to_quantize, high_precision_dtype).


---------

+++ code/evaluation/logical_error_rate.py
@@ -200,7 +200,7 @@ def _ort_quantize_int8(fp32_onnx_path: str, output_path: str, calib_dets: "np.nd
     class _DetCalibReader(CalibrationDataReader):
 
         def __init__(self, data):
-            self._rows = [{"dets": data[i:i + 1].astype("float32")} for i in range(len(data))]
+            self._rows = [{"dets": data[i:i + 1]} for i in range(len(data))]
             self._iter = iter(self._rows)
 
         def get_next(self):
@@ -1202,7 +1202,7 @@ def run_inference_and_decode_pre_decoder_memory(model, device, dist, cfg) -> dic
                             mq.quantize(
                                 onnx_path=fp32_onnx_path,
                                 quantize_mode=quant_format,
-                                calibration_data={"dets": calib_dets.astype("float32")},
+                                calibration_data={"dets": calib_dets},
                                 output_path=onnx_path,
                                 **quant_kwargs,
                             )
